+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### PR links (https://github.com/wesleychipeco/trifecta-website-v3/pulls?q=is%3Apr+is%3Aclosed)
 
+### [4.0.0] - 2025-03-26
+
 ### [PR-93] - 2025-03-03
 
 - New commissioner action: add leagueId and teamIds for a new sport via UI instead of manually through DB
@@ -13,9 +15,6 @@
 ### [PR-91] - 2025-02-20
 
 - Throttle API requests to Fantrax for /rosters and /player-stats scraping
-
-### [4.0.0] - 2025-02-15
-
 - Include changes to Google doc about httpd server updates for new endpoints
 
 ### Changed